Add server check and toggle shutdown button
This commit introduces a check, if the raspberry_server.py is running or not (also allowing for future reference to deliver files). If the script was not started on (re-)boot, the shutdown button will not be present.
Description on how to start the server was updated.
Check if server is online is done via a small ping-picture as cross-site scripting is not allowed.

gui_sv is the raspberry server folder, where all files for :8001 requests can be stored.