Introduce controller selection
Introduce an option to select a hardware controller.

Updates:
- Update of config file -> needs to be added manually to old files
- index.html: to identify the webgui clearly as a controller; remove invert of axis. This is now done in the joy_stick_parser.
- joystick_parser_node: config dicts for different controllers (currently Logitech F710 and X-Box One)

Introduced sensitivity on controller to allow for a certain threshold. Tests have shown, that even all sticks are released, a controller might send a very low value. The joy_node has a similar "deadzone"-parameter, but this parameter is not used on purpose to allow for different sensitivity based on different controlling methods.